% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creator.R
\name{creator}
\alias{creator}
\title{Create Distributed Coding Tasks}
\usage{
creator(input_files_dir = NULL, path_to_form = NULL,
  output_files_dir = NULL, n_per_worker = 10, n_per_story = 3,
  seed = 31415)
}
\arguments{
\item{input_files_dir}{path to directory containing text files that need to be coded; required}

\item{path_to_form}{path to html form; required}

\item{output_files_dir}{path to directory containing text files that need to be coded; required}

\item{n_per_worker}{maximum no. of stories that you want each worker to code}

\item{n_per_story}{no. of times you want each story coded}

\item{seed}{seed for the random number generator}
}
\description{
This function calculates the edit distance between OCR'd and human transcribed file.
The function by default prints the status of the task you are trying to delete. It will show up as 'deleted' if successful
}
\examples{
\dontrun{
creator(input_files_dir="path_to_input_files_dir", path_to_form="path_tp_html_form", 
		   output_files_dir="path_to_output_files_dir", n_per_worker=20, n_per_story=3)
}
}

