% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outsource.R
\docType{package}
\name{outsource-package}
\alias{outsource}
\alias{outsource-package}
\title{Outsource: Create Distributed Coding Tasks More Easily}
\description{
More easily create distributed Human Intelligence Coding Tasks
}
\author{
Gaurav Sood
}

