% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outsource.r
\name{outsource}
\alias{outsource}
\title{This function evaluates the script at \code{file_path} and binds what this script returns to \code{.} variable and then evaluates \code{what_to_do} with this variable added to the evaluation environment.}
\usage{
outsource(
  file_path,
  what_to_do,
  return_if_error = NULL,
  return_if_no_file = return_if_error,
  bind_file_path = FALSE
)
}
\arguments{
\item{file_path}{Path to R script}

\item{what_to_do}{Object to be evaluated (see \code{expr} argument in \code{eval})}

\item{return_if_error}{What to return if error occurred}

\item{return_if_no_file}{What to return if there is no script}

\item{bind_file_path}{Whether to also bind \code{file_path} as \code{.path}}
}
\value{
Evaluation result
}
\description{
This function evaluates the script at \code{file_path} and binds what this script returns to \code{.} variable and then evaluates \code{what_to_do} with this variable added to the evaluation environment.
}
